Change Makefile default target to build
The current Makefile has a default target of install. This will install
the plugin into you users' terraform directory. This is a little
uncommon, as usually `make` does not change anything outside the repo.
That is what `make install` is used for.

This commit changes the default make target from `install` to `build`.

Signed-off-by: Jeroen Simonetti <jeroen@simonetti.nl>